Add property to control Viewable evaluation.
see: #34

At the moment, Krazo assumes that every response entity
has to be a view which can be processed by a ViewEngine.
In case there is e.g. some custom exception mapper who wants to
return a plain text message, the Request stops with an exception.

To avoid this, Krazo can now be configured to either use this strict
evaluation or some less strict which enables Krazo to return entities
which are not views. To enable this, the config
'org.eclipse.krazo.strictViewResolution' has to be set to false.

Signed-off-by: Erdle, Tobias <tobias.erdle@innoq.com>