Enable ConverterRegistry to register custom converters
see: #118

Since now a user is able to register its own MvcConverter in the ConverterRegistry.
This should enable them to provide more specific converters for special cases. To
ensure the custom converters are loaded before the built-in ones, the MvcConverter
is now prioritizable and custom implementations can choose in which order they
should be loaded.

Signed-off-by: Erdle, Tobias <tobias.erdle@innoq.com>