Update test dependencies in core module
see: #115

The test dependencies within the core module are up-to-date now. Unfortunately,
Weld changed its behavior and wanted to load servlet related beans which
aren't necessary in this test. As a result, it didn't work anymore. As I don't see
any reason to use CDI for testing the annotation hierachy, I removed CDI and the test
passes again.

Signed-off-by: Erdle, Tobias <tobias.erdle@innoq.com>