Add testsuite task for Websphere Liberty
see: #155

To ensure Krazo works on Liberty as well as on TomEE in any case, a new task for Travis is added.
The reason for this is, that Liberty behaves different e.g. when setting classes in the JAX-RS
application which leads to a lot of bug reports and questions.

Signed-off-by: Erdle, Tobias <tobias.erdle@innoq.com>