Activate hidden method filtering by toggle (#207)
* Activate hidden method filtering by toggle

At the moment, the HiddenMethodFilter is active all the time and checks every
request for possibly changed methods. I could imagine that this is not
wanted by every user, because they maybe want to use other mechanisms to
determine the HTTP method or just use the HTML Form standards. To ensure
that this feature doesn't cause trouble to anyone who isn't aware of the feature,
now a toggle is added to the KrazoProperties and the filter can be activated in the JAX-RS
application.

Signed-off-by: Tobias Erdle <tobias.erdle@innoq.com>

* Add HiddenMethodFilter to CDI extension

Signed-off-by: Tobias Erdle <tobias.erdle@innoq.com>