Remove CXF from testsuite and TCK
Be aware that the Liberty profile is only commented out, as we will
support this server again, as soon as they use RESTEasy. Because of
this, the archive creator is adapted to use RESTEasy, so later we only
need to remove the comments from the profiles and adapt the dependecies
within the pom.xml.

Signed-off-by: Erdle, Tobias <tobi.erdle@gmail.com>