Replace test running inside container with remote variant
As it seems that the test was flaky because of the namespace change from
javax -> jakarta and a update of Arquillian and HTMLUnit didn't help, I
decided to replace the test with a similar one which runs outside the
container. This fixes the build errors on Glassfish and gives us the
same information.

fixes: #276