[#233] Reenabled Thymeleaf extension and adapted to changes in Thymeleaf
This reenables the Thymeleaf extension by using its 3.1.0.M1 milestone release (which uses the jakarta namespace).
To adapt to changes in Thymeleaf, CDIWebContext had to be changed to use the jakarta namespace and the new IWebExchange.
Since `JakartaServletWebApplication` has been established to wrap ServletContext in Thymeleaf, there needed to be a producer to inject it into both `DefaultTemplateEngineProducer` and `ThymeleafViewEngine`.
But because `JakartaServletWebApplication` is a final class, a wrapper for it had to be written in order to use CDI.