Use Jakarta EE 10 APIs and implementation version
RESTEasy renamed some of their artifacts we used before, so they
had to be updated. Those new artifacts seem to be the correct ones.

In all other cases it was enough to increase the version numbers.
Arquillian adapters etc. were not upgraded as they are part of a
separate ticket and may need more attention.

fixes: 318