Rename archetype to generic name
At the moment, we released the archetypes with a hint which
Jakarta EE version they're targeting. This leads to a confusing
versioning scheme, as there is the archetype for EE 9 beginning with
version 2.0.0 and so on. To avoid this, the archetype's Jakarta EE
version is now removed and shall be indicated by the Jakarta MVC and
Krazo version used.

see: #336