Fix annotation check for controller methods in RESTEasy
While fixing #334 (PR #335) we introduced a small bug during the check
if a method needs to be processed by the general validator or not.
Because of this bug, the implementation checked only if a method is
annotated with @Controller, but not if this is the case for the whole
class. The result was Krazo not handling invalid input correctly.

This fix adds the ability to check if the method OR the declaring class
is annotated with @Controller, so we ensure we really have something
that needs to be handled by MVC.

This fix needs to be backported to Krazo 3.0.x.

fixes: #342